* vm_dump.c (rb_print_backtrace): return value of libexec's backtrace   is size_t, so simply cast as int.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@50553 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ vm_dump.c
@@ -690,7 +690,7 @@ rb_print_backtrace(void)
 #if HAVE_BACKTRACE
 #define MAX_NATIVE_TRACE 1024
     static void *trace[MAX_NATIVE_TRACE];
-    int n = backtrace(trace, MAX_NATIVE_TRACE);
+    int n = (int)backtrace(trace, MAX_NATIVE_TRACE);
 #if defined(USE_ELF) && defined(HAVE_DLADDR)
     rb_dump_backtrace_with_lines(n, trace);
 #else
